fix: move common methods
