+++ pkg/client/common/channel.go
@@ -1,4 +1,4 @@
-package tests
+package common
 
 import "sync/atomic"
